Registry pattern for training setups
