+++ finetune.py
@@ -1,16 +1,29 @@
-import torch
-import wandb, os, json
-import time
-import numpy as np
-from torch.amp import autocast
 import gc
 import argparse
+import time
+import os
+import json
+
+import wandb
+import numpy as np
+import torch
+from torch.amp import autocast
+from torch.optim.lr_scheduler import ChainedScheduler
 
 from utils.muon_optimizer import Muon
 from subject.dataset import load_subjects
 from evaluation.neuroprobe_tasks import FrozenModelEvaluation_SS_SM
-from training_setup.training_config import log, update_dir_name, update_random_seed, parse_config_from_args, get_default_config, parse_subject_trials_from_config, convert_dtypes, unconvert_dtypes
-from torch.optim.lr_scheduler import ChainedScheduler
+from training_setup.registry import resolve
+from training_setup.training_config import (
+    log, 
+    update_dir_name, 
+    update_random_seed, 
+    parse_config_from_args, 
+    get_default_config, 
+    parse_subject_trials_from_config, 
+    convert_dtypes, 
+    unconvert_dtypes
+)
 
 ### PARSE ARGUMENTS ###
 
@@ -157,14 +170,8 @@ all_subjects = load_subjects(config['training']['train_subject_trials'],
 ### LOADING TRAINING SETUP ###
 
 # Import the training setup class dynamically based on config
-training_setup_name = config["training"]["setup_name"].lower() # if this is X, the filename should be trianing_setup/X.py and the class name should be XTrainingSetup
-try:
-    setup_module = __import__(f'training_setup.{training_setup_name}', fromlist=[training_setup_name])
-    setup_class = getattr(setup_module, training_setup_name)
-    training_setup = setup_class(all_subjects, config, verbose=True)
-except (ImportError, AttributeError) as e:
-    print(f"Could not load training setup '{config['training']['setup_name']}'. Are you sure the filename and the class name are the same and correspond to the parameter? Error: {str(e)}")
-    exit()
+training_setup_name = config["training"]["setup_name"] # Name in registry
+training_setup = resolve(training_setup_name, all_subjects=all_subjects, config=config, verbose=True)
 
 # Save a copy of the training setup file for reproducibility
 import shutil
