+++ training_setup/training_setup.py
@@ -11,22 +11,18 @@ import json
 import torch
 from torch.utils.data import DataLoader, ConcatDataset
 
+from subject.dataset import SubjectTrialDataset, PreprocessCollator, SubjectBatchSampler
+from training_setup.training_config import log, convert_dtypes
+from model.preprocessing.laplacian_rereferencing import laplacian_rereference_batch
 from model.electrode_embedding import (
     ElectrodeEmbedding_Learned,
     ElectrodeEmbedding_NoisyCoordinate,
     ElectrodeEmbedding_Learned_CoordinateInit,
     ElectrodeEmbedding_Zero,
 )
-from model.preprocessing.laplacian_rereferencing import laplacian_rereference_batch
-
-from subject.dataset import SubjectTrialDataset, PreprocessCollator, SubjectBatchSampler
-
-from training_setup.training_config import (
-    log,
-    convert_dtypes,
-)
 
 RUN_DIR = 'runs/data'
+
 class TrainingSetup:
     '''
     Interface for training setups.
